requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn==0.30.6

# Database
sqlalchemy==2.0.35
pymysql==1.1.1

# Data & ML
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.14.1

# Environment
python-dotenv==1.0.1